+++ tests/testthat/test-adv-dplyr.R
@@ -1,12 +1,14 @@
 context("adv-dplyr")
 
-
 library(maptools)
 library(dplyr)
 data(wrld_simpl)
 poly1 <- wrld_simpl
 line1 <- as(wrld_simpl, "SpatialLinesDataFrame")
 point1 <- as(line1, "SpatialPointsDataFrame")
+expect_silent(g <- wrld_simpl  %>% group_by(REGION)  %>% summarize(alon = mean(LON), mxlat = max(LAT), mxarea = max(AREA)))
+
+
 #mpoint1 <- SpatialMultiPointsDataFrame(lapply(split(line1, seq(nrow(line1))), function(y) coordinates(as(y, "SpatialPoints"))), 
 #                                       as.data.frame(line1))
 # filter() (and slice())